Added a script to filter the probe data out.
Added get_probe.py, which is a script to separate the training data into
the pure training data and the probe data.  It's kind of slow, and depends
on how you named your files/where you put them.  It also relies on having
already gotten user files, though it's easily modifiable to only work with
movies, if you so wish it.  Seems to work for the movies, and I'll see if
it works for the user files when I wake up.