Added a function to compile probe data by users.
The probe.readProbeDataByUser() is similar to the
probe.readProbeData(), except that it groups by users rather than
by movies. The saved file format is also a little different.

The header is:
MAGIC_ID, VERSION, num users
(The version is used to denote what code was used to generate the
 database. If any code that is used to save the probe data
 is changed, please increment the USER_PROBE_VERSION in probe.cpp)

The records then are:
userId numRatings
movieId rating
movieId rating
....