Added a probe error function on ordering.
The probe.runProbeOrder runs a error calculation on an
OrderingAlgorithm. The error function is as follows:

For each user, a user error is calculated. For each movie in
the test set, it is compared against every movie in the training
set. If the movies are rated the same, the pair is thrown out
as a test candidate. Otherwise, an error is given if the algorithm
predicted the wrong order. A similar procedure is done for the
pairs within the test set. The user error is then the number of
errors divided by the number of test candidates. The overall error
is the average user error over all users.