Removed correlation row sorting
Removed sorting of correlation rows as it was taking too much time.
Instead the correlation rows will be unsorted, and every relevant
correlation coefficient will be used. Because of this, the movieId
doesn't need to be attached to its correlation (it is implied by
the index into the array), so it was removed to save space.